+++ packages/agents-core/src/utils/index.ts
@@ -9,5 +9,5 @@ export * from './logger';
 export * from './mcp-client';
 export * from './schema-conversion';
 export * from './service-token-auth';
-export * from './third-party-check';
+export * from './third-party-mcp-servers';
 export * from './tracer-factory';
